requirements.txt
beautifulsoup4==4.12.3
ffpyplayer==4.5.1
fonttools==4.43.1
googlesearch_python==1.2.3
keyboard==0.13.5
moviepy==1.0.3
opencv_python==4.8.1.78
Pillow
psutil==5.9.5
pygame==2.5.2
pymediainfo==6.0.1
pytube==15.0.0
pyvidplayer2==0.9.11
pywin32==306
scapy==2.5.0
screeninfo==0.8.1
selenium==4.21.0
youtube_search_python==1.6.6
